Deployment.yaml fix, soloVolume concept added
The `deployment.yml` was invalid due to indentation changes made from an
earlier commit. These changes have been confirmed via `helm install
--dry-run --debug plex kube-plex-hector/charts/kube-plex/` to ensure
that it compiles correctly.

The concept of a `soloVolume` and configurations have been added where
if we wanted to mount all of the `transcode`, `data`, `config` files and
    values, they could be mounted on a single PV/PVC.